Move uniquePoints into hulls
