+++ packages/modeling/src/operations/hulls/hullGeom3.js
@@ -4,6 +4,7 @@ const geom3 = require('../../geometries/geom3')
 const poly3 = require('../../geometries/poly3')
 
 const quickhull = require('./quickhull')
+const toUniquePoints = require('./toUniquePoints')
 
 /*
  * Create a convex hull of the given geometries (geom3).
@@ -16,26 +17,12 @@ const hullGeom3 = (...geometries) => {
   if (geometries.length === 1) return geometries[0]
 
   // extract the unique vertices from the geometries
-  const uniquevertices = []
-  const found = new Map()
-  for (let g = 0; g < geometries.length; ++g) {
-    const polygons = geom3.toPolygons(geometries[g])
-    for (let p = 0; p < polygons.length; ++p) {
-      const vertices = polygons[p].vertices
-      for (let v = 0; v < vertices.length; ++v) {
-        const id = `${vertices[v]}`
-        if (found.has(id)) continue
-        uniquevertices.push(vertices[v])
-        found.set(id, true)
-      }
-    }
-  }
-  found.clear()
+  const unique = toUniquePoints(geometries)
 
-  const faces = quickhull(uniquevertices, { skipTriangulation: true })
+  const faces = quickhull(unique, { skipTriangulation: true })
 
   const polygons = faces.map((face) => {
-    const vertices = face.map((index) => uniquevertices[index])
+    const vertices = face.map((index) => unique[index])
     return poly3.create(vertices)
   })
