+++ packages/modeling/src/operations/hulls/toUniquePoints.js
@@ -1,11 +1,11 @@
-const geom2 = require("../geometries/geom2");
-const geom3 = require("../geometries/geom3");
-const path2 = require("../geometries/path2");
+const geom2 = require('../../geometries/geom2')
+const geom3 = require('../../geometries/geom3')
+const path2 = require('../../geometries/path2')
 
 /*
  * Return the unique vertices of a geometry
  */
-const uniquePoints = (geometries) => {
+const toUniquePoints = (geometries) => {
   const found = new Set()
   const uniquePoints = []
 
@@ -21,7 +21,8 @@ const uniquePoints = (geometries) => {
     if (geom2.isA(geometry)) {
       geom2.toPoints(geometry).forEach(addPoint)
     } else if (geom3.isA(geometry)) {
-      geom3.toPoints(geometry).forEach(addPoint)
+      // points are grouped by polygon
+      geom3.toPoints(geometry).forEach((points) => points.forEach(addPoint))
     } else if (path2.isA(geometry)) {
       path2.toPoints(geometry).forEach(addPoint)
     }
@@ -30,4 +31,4 @@ const uniquePoints = (geometries) => {
   return uniquePoints
 }
 
-module.exports = uniquePoints
+module.exports = toUniquePoints
