+++ packages/modeling/src/utils/index.d.ts
@@ -5,7 +5,6 @@ export { default as fnNumberSort } from './fnNumberSort'
 export { default as insertSorted } from './insertSorted'
 export { default as radiusToSegments } from './radiusToSegments'
 export { default as radToDeg } from './radToDeg'
-export { default as uniquePoints } from './uniquePoints'
 export * from './trigonometry'
 
 export as namespace utils
