+++ packages/modeling/src/utils/index.js
@@ -13,6 +13,5 @@ module.exports = {
   insertSorted: require('./insertSorted'),
   radiusToSegments: require('./radiusToSegments'),
   radToDeg: require('./radToDeg'),
-  sin: require('./trigonometry').sin,
-  uniquePoints: require('./uniquePoints')
+  sin: require('./trigonometry').sin
 }
